phylo_place_agglom/scripts/requirements.txt
ete3
biopython
